Split lexer string tests
