+++ src/lexer/tests.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod string_literals;
+
 /// Tokenise and return just token variants, filtering out Newline/EOF.
 fn toks(src: &str) -> Vec<Token> {
     tokenize(src, 0)
@@ -85,100 +87,6 @@ fn hex_binary_octal() {
             Token::IntLiteral(0b1010),
             Token::IntLiteral(0o777),
             Token::IntLiteral(0xDEAD),
-        ]
-    );
-}
-
-#[test]
-fn plain_string() {
-    assert_eq!(
-        toks(r#""hello world""#),
-        vec![Token::StringLiteral("hello world".into())]
-    );
-}
-
-#[test]
-fn escape_sequences() {
-    assert_eq!(
-        toks(r#""\n\t\\\"""#),
-        vec![Token::StringLiteral("\n\t\\\"".into())]
-    );
-}
-
-#[test]
-fn hex_escape() {
-    assert_eq!(toks(r#""\x41""#), vec![Token::StringLiteral("A".into())]);
-}
-
-#[test]
-fn null_escape() {
-    assert_eq!(toks(r#""\0""#), vec![Token::StringLiteral("\0".into())]);
-}
-
-#[test]
-fn escaped_dollar_no_interpolation() {
-    assert_eq!(
-        toks(r#""\${not interpolated}""#),
-        vec![Token::StringLiteral("${not interpolated}".into())]
-    );
-}
-
-#[test]
-fn string_interpolation_simple() {
-    assert_eq!(
-        toks(r#""hello ${name}!""#),
-        vec![
-            Token::StringChunk("hello ".into()),
-            Token::InterpolationStart,
-            Token::Identifier("name".into()),
-            Token::InterpolationEnd,
-            Token::StringChunk("!".into()),
-        ]
-    );
-}
-
-#[test]
-fn string_interpolation_expr() {
-    assert_eq!(
-        toks(r#""result = ${a + b}""#),
-        vec![
-            Token::StringChunk("result = ".into()),
-            Token::InterpolationStart,
-            Token::Identifier("a".into()),
-            Token::Plus,
-            Token::Identifier("b".into()),
-            Token::InterpolationEnd,
-        ]
-    );
-}
-
-#[test]
-fn string_interpolation_call() {
-    assert_eq!(
-        toks(r#""${f(x)}""#),
-        vec![
-            Token::InterpolationStart,
-            Token::Identifier("f".into()),
-            Token::LParen,
-            Token::Identifier("x".into()),
-            Token::RParen,
-            Token::InterpolationEnd,
-        ]
-    );
-}
-
-#[test]
-fn string_interpolation_multiple() {
-    assert_eq!(
-        toks(r#""${a} and ${b}""#),
-        vec![
-            Token::InterpolationStart,
-            Token::Identifier("a".into()),
-            Token::InterpolationEnd,
-            Token::StringChunk(" and ".into()),
-            Token::InterpolationStart,
-            Token::Identifier("b".into()),
-            Token::InterpolationEnd,
         ]
     );
 }
